Add an experimental flag to the preferences pane, reorganize the in-game book based data representation implementation, use Objects.hash instead of manually calculating hash codes in DataModels, update Config.md to reflect the changes
