+++ src/main/java/com/redcreator37/playermanagement/Commands/PlayerCommands/RegisterId.java
@@ -5,7 +5,7 @@ import com.redcreator37.playermanagement.Commands.PlayerCommand;
 import com.redcreator37.playermanagement.DataModels.PlayerTag;
 import com.redcreator37.playermanagement.DataModels.ServerPlayer;
 import com.redcreator37.playermanagement.Localization;
-import com.redcreator37.playermanagement.PlayerCard;
+import com.redcreator37.playermanagement.IdHandling.PlayerCard;
 import com.redcreator37.playermanagement.PlayerManagement;
 import com.redcreator37.playermanagement.PlayerRoutines;
 import org.bukkit.ChatColor;
