+++ src/main/java/com/redcreator37/playermanagement/DataModels/ServerPlayer.java
@@ -136,18 +136,8 @@ public class ServerPlayer {
      */
     @Override
     public int hashCode() {
-        int result = 17;
-        result *= 37 + getId();
-        result *= 37 + Objects.hashCode(getUsername());
-        result *= 37 + Objects.hashCode(getUuid());
-        result *= 37 + Objects.hashCode(getName());
-        result *= 37 + Objects.hashCode(getJoinDate());
-        result *= 37 + Objects.hashCode(getJob());
-        result *= 37 + Objects.hashCode(getCompany());
-        result *= 37 + Objects.hashCode(getNotes());
-        result *= 37 + getPunishments();
-        result *= 37 + getPunishments();
-        return result;
+        return Objects.hash(getId(), getUsername(), getUuid(), getName(),
+                getJoinDate(), getJob(), getCompany(), getNotes());
     }
 
 }
